feat(server): Build AI framework. (#29)
* feat(server): Build AI framework.


* update(server): Refine AI framework.


* update(server): Add example to use PlanAgent.


* feat(server): Add ES for full-text search.


* update(server): Refine plan agent and trip creation.


* update(server): Refine plan agent workflow and prompt


* refactor(server): Tidy the codes.


---------

+++ internal/data/ent/schema/trip.go
@@ -22,7 +22,7 @@ func (Trip) Fields() []ent.Field {
 		field.String("description").MaxLen(255),
 		field.Time("start_date"),
 		field.Time("end_date"),
-		field.Int("days").NonNegative(),
+		field.Int8("days").NonNegative(),
 	}
 }
